+++ examples/admin-saml-groups/py/__main__.py
@@ -1,0 +1,9 @@
+"""A Python Pulumi program"""
+
+import pulumi
+import pulumi_splunk as splunk
+
+saml_group = splunk.AdminSamlGroups("demo-py-group", roles=[
+    "admin",
+    "power",
+])