+++ examples/examples_dotnet_test.go
@@ -1,0 +1,22 @@
+// Copyright 2016-2017, Pulumi Corporation.  All rights reserved.
+// +build dotnet all
+
+package examples
+
+import (
+	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
+)
+
+func getCSBaseOptions(t *testing.T) integration.ProgramTestOptions {
+	envRegion := getEnvRegion(t)
+	base := getBaseOptions()
+	csharpBase := base.With(integration.ProgramTestOptions{
+		Dependencies: []string{
+			"Pulumi.Splunk",
+		},
+	})
+
+	return csharpBase
+}