+++ provider/resources.go
@@ -12,26 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package xyz
+package splunk
 
 import (
 	"unicode"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/pulumi/pulumi-terraform-bridge/v2/pkg/tfbridge"
-	shim "github.com/pulumi/pulumi-terraform-bridge/v2/pkg/tfshim"
 	shimv1 "github.com/pulumi/pulumi-terraform-bridge/v2/pkg/tfshim/sdk-v1"
-	"github.com/pulumi/pulumi/sdk/v2/go/common/resource"
 	"github.com/pulumi/pulumi/sdk/v2/go/common/tokens"
-	"github.com/terraform-providers/terraform-provider-xyz/xyz"
+	"github.com/splunk/terraform-provider-splunk/splunk"
 )
 
 // all of the token components used below.
 const (
 	// packages:
-	mainPkg = "xyz"
+	mainPkg = "splunk"
 	// modules:
-	mainMod = "index" // the y module
+	mainMod = "index" // the main module
 )
 
 // makeMember manufactures a type token for the package and the given module and type.
@@ -47,10 +45,10 @@ func makeType(mod string, typ string) tokens.Type {
 // makeDataSource manufactures a standard resource token given a module and resource name.  It
 // automatically uses the main package and names the file by simply lower casing the data source's
 // first character.
-func makeDataSource(mod string, res string) tokens.ModuleMember {
-	fn := string(unicode.ToLower(rune(res[0]))) + res[1:]
-	return makeMember(mod+"/"+fn, res)
-}
+//func makeDataSource(mod string, res string) tokens.ModuleMember {
+//	fn := string(unicode.ToLower(rune(res[0]))) + res[1:]
+//	return makeMember(mod+"/"+fn, res)
+//}
 
 // makeResource manufactures a standard resource token given a module and resource name.  It
 // automatically uses the main package and names the file by simply lower casing the resource's
@@ -60,93 +58,155 @@ func makeResource(mod string, res string) tokens.Type {
 	return makeType(mod+"/"+fn, res)
 }
 
-// boolRef returns a reference to the bool argument.
-func boolRef(b bool) *bool {
-	return &b
-}
-
-// stringValue gets a string value from a property map if present, else ""
-func stringValue(vars resource.PropertyMap, prop resource.PropertyKey) string {
-	val, ok := vars[prop]
-	if ok && val.IsString() {
-		return val.StringValue()
-	}
-	return ""
-}
-
-// preConfigureCallback is called before the providerConfigure function of the underlying provider.
-// It should validate that the provider can be configured, and provide actionable errors in the case
-// it cannot be. Configuration variables can be read from `vars` using the `stringValue` function -
-// for example `stringValue(vars, "accessKey")`.
-func preConfigureCallback(vars resource.PropertyMap, c shim.ResourceConfig) error {
-	return nil
-}
-
-// managedByPulumi is a default used for some managed resources, in the absence of something more meaningful.
-var managedByPulumi = &tfbridge.DefaultInfo{Value: "Managed by Pulumi"}
-
 // Provider returns additional overlaid schema and metadata associated with the provider..
 func Provider() tfbridge.ProviderInfo {
-	// Instantiate the Terraform provider
-	p := shimv1.NewProvider(xyz.Provider().(*schema.Provider))
+	p := shimv1.NewProvider(splunk.Provider().(*schema.Provider))
 
-	// Create a Pulumi provider mapping
 	prov := tfbridge.ProviderInfo{
 		P:           p,
-		Name:        "xyz",
-		Description: "A Pulumi package for creating and managing xyz cloud resources.",
-		Keywords:    []string{"pulumi", "xyz"},
+		Name:        "splunk",
+		Description: "A Pulumi package for creating and managing splunk cloud resources.",
+		Keywords:    []string{"pulumi", "splunk"},
 		License:     "Apache-2.0",
 		Homepage:    "https://pulumi.io",
-		Repository:  "https://github.com/pulumi/pulumi-xyz",
-		Config:      map[string]*tfbridge.SchemaInfo{
-			// Add any required configuration here, or remove the example below if
-			// no additional points are required.
-			// "region": {
-			// 	Type: makeType("region", "Region"),
-			// 	Default: &tfbridge.DefaultInfo{
-			// 		EnvVars: []string{"AWS_REGION", "AWS_DEFAULT_REGION"},
-			// 	},
-			// },
+		Repository:  "https://github.com/pulumi/pulumi-splunk",
+		GitHubOrg:   "splunk",
+		Config:      map[string]*tfbridge.SchemaInfo{},
+		Resources: map[string]*tfbridge.ResourceInfo{
+			"splunk_admin_saml_groups": {
+				Tok: makeResource(mainMod, "AdminSamlGroups"),
+				Docs: &tfbridge.DocInfo{
+					Source: "admin_saml_groups.md",
+				},
+			},
+			"splunk_apps_local": {
+				Tok: makeResource(mainMod, "AppsLocal"),
+				Docs: &tfbridge.DocInfo{
+					Source: "apps_local.md",
+				},
+			},
+			"splunk_authentication_users": {
+				Tok: makeResource(mainMod, "AuthenticationUsers"),
+				Docs: &tfbridge.DocInfo{
+					Source: "authentication_users.md",
+				},
+			},
+			"splunk_authorization_roles": {
+				Tok: makeResource(mainMod, "AuthorizationRoles"),
+				Docs: &tfbridge.DocInfo{
+					Source: "authorization_roles.md",
+				},
+			},
+			"splunk_global_http_event_collector": {
+				Tok: makeResource(mainMod, "GlobalHttpEventCollector"),
+				Docs: &tfbridge.DocInfo{
+					Source: "global_http_event_collector.md",
+				},
+			},
+			"splunk_inputs_http_event_collector": {
+				Tok: makeResource(mainMod, "InputsHttpEventCollector"),
+				Docs: &tfbridge.DocInfo{
+					Source: "inputs_http_event_collector.md",
+				},
+			},
+			"splunk_inputs_script": {
+				Tok: makeResource(mainMod, "InputsScript"),
+				Docs: &tfbridge.DocInfo{
+					Source: "inputs_script.md",
+				},
+			},
+			"splunk_inputs_monitor": {
+				Tok: makeResource(mainMod, "InputsMonitor"),
+				Docs: &tfbridge.DocInfo{
+					Source: "inputs_monitor.md",
+				},
+			},
+			"splunk_inputs_udp": {
+				Tok: makeResource(mainMod, "InputsUdp"),
+				Docs: &tfbridge.DocInfo{
+					Source: "inputs_udp.md",
+				},
+			},
+			"splunk_inputs_tcp_raw": {
+				Tok: makeResource(mainMod, "InputsTcpRaw"),
+				Docs: &tfbridge.DocInfo{
+					Source: "inputs_tcp_raw.md",
+				},
+			},
+			"splunk_inputs_tcp_cooked": {
+				Tok: makeResource(mainMod, "InputsTcpCooked"),
+				Docs: &tfbridge.DocInfo{
+					Source: "inputs_tcp_cooked.md",
+				},
+			},
+			"splunk_inputs_tcp_splunk_tcp_token": {
+				Tok: makeResource(mainMod, "InputsTcpSplunkTcpToken"),
+				Docs: &tfbridge.DocInfo{
+					Source: "inputs_tcp_splunktcptoken.md",
+				},
+			},
+			"splunk_inputs_tcp_ssl": {
+				Tok: makeResource(mainMod, "InputsTcpSsl"),
+				Docs: &tfbridge.DocInfo{
+					Source: "inputs_tcp_ssl.md",
+				},
+			},
+			"splunk_outputs_tcp_default": {
+				Tok: makeResource(mainMod, "OutputsTcpDefault"),
+				Docs: &tfbridge.DocInfo{
+					Source: "outputs_tcp_default.md",
+				},
+			},
+			"splunk_outputs_tcp_server": {
+				Tok: makeResource(mainMod, "OutputsTcpServer"),
+				Docs: &tfbridge.DocInfo{
+					Source: "outputs_tcp_server.md",
+				},
+			},
+			"splunk_outputs_tcp_group": {
+				Tok: makeResource(mainMod, "OutputsTcpGroup"),
+				Docs: &tfbridge.DocInfo{
+					Source: "outputs_tcp_group.md",
+				},
+			},
+			"splunk_outputs_tcp_syslog": {
+				Tok: makeResource(mainMod, "OutputsTcpSyslog"),
+				Docs: &tfbridge.DocInfo{
+					Source: "outputs_tcp_syslog.md",
+				},
+			},
+			"splunk_saved_searches": {
+				Tok: makeResource(mainMod, "SavedSearches"),
+				Docs: &tfbridge.DocInfo{
+					Source: "saved_searches.md",
+				},
+			},
+			"splunk_indexes": {
+				Tok: makeResource(mainMod, "Indexes"),
+				Docs: &tfbridge.DocInfo{
+					Source: "indexes.md",
+				},
+			},
+			"splunk_configs_conf": {
+				Tok: makeResource(mainMod, "ConfigsConf"),
+				Docs: &tfbridge.DocInfo{
+					Source: "configs_conf.md",
+				},
+			},
 		},
-		PreConfigureCallback: preConfigureCallback,
-		Resources:            map[string]*tfbridge.ResourceInfo{
-			// Map each resource in the Terraform provider to a Pulumi type. Two examples
-			// are below - the single line form is the common case. The multi-line form is
-			// needed only if you wish to override types or other default options.
-			//
-			// "aws_iam_role": {Tok: makeResource(mainMod, "IamRole")}
-			//
-			// "aws_acm_certificate": {
-			// 	Tok: makeResource(mainMod, "Certificate"),
-			// 	Fields: map[string]*tfbridge.SchemaInfo{
-			// 		"tags": {Type: makeType(mainPkg, "Tags")},
-			// 	},
-			// },
-		},
-		DataSources: map[string]*tfbridge.DataSourceInfo{
-			// Map each resource in the Terraform provider to a Pulumi function. An example
-			// is below.
-			// "aws_ami": {Tok: makeDataSource(mainMod, "getAmi")},
-		},
+		DataSources: map[string]*tfbridge.DataSourceInfo{},
 		JavaScript: &tfbridge.JavaScriptInfo{
-			// List any npm dependencies and their versions
 			Dependencies: map[string]string{
-				"@pulumi/pulumi": "^2.0.0",
+				"@pulumi/pulumi": "^2.15.0",
 			},
 			DevDependencies: map[string]string{
 				"@types/node": "^8.0.25", // so we can access strongly typed node definitions.
 				"@types/mime": "^2.0.0",
 			},
-			// See the documentation for tfbridge.OverlayInfo for how to lay out this
-			// section, or refer to the AWS provider. Delete this section if there are
-			// no overlay files.
-			//Overlay: &tfbridge.OverlayInfo{},
 		},
 		Python: &tfbridge.PythonInfo{
-			// List any Python dependencies and their version ranges
 			Requires: map[string]string{
-				"pulumi": ">=2.9.0,<3.0.0",
+				"pulumi": ">=2.15.0,<3.0.0",
 			},
 		},
 		CSharp: &tfbridge.CSharpInfo{