+++ CHANGELOG.md
@@ -2,5 +2,6 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+* Initial creation of the provider against Splunk Terraform Provider v1.3.7
 
 ---